Fix pointer provenance in FcPatternObjectInsertElt()
When FcPatternElt and FcPattern are not part of the same underlying
memory allocation, creating a pointer to FcPatternElt derived from
the FcPattern allocation is undefined behaviour. While this works on
conventional architectures, it causes issues for architectures that
track pointer bounds at runtime (e.g. CHERI-enabled architectures such
as Arm's Morello). To avoid this, we can add a `FcIsEncodedOffset()`
check to `FcPatternElts()` and only use `FcOffsetMember()` in the cases
where elts_offset points to the same allocation. This also allows changing
the structure member to be a FcPatternElt * rather than an intptr_t.

Without this change I get a runtime crash when trying to run the poppler
unit test check_qt5_fonts on CHERI-RISC-V CheriBSD, and with it I can
successfully run all the poppler unit test them.